v4.0 - Route restructure - move Campgrounds and Reviews routes out to separate files - Ensure dependencies and middleware are moved as well - Make sure path of dependencies and middleware is correct
